Fix signature issues

+++ lcms2.net/types/type_handlers/CurveHandler.cs
@@ -43,7 +43,7 @@ public class CurveHandler : TagTypeHandler
                 if (count > 0x7FFF)
                     return null; // This is to prevent bad guys for doing bad things.
 
-                newGamma = ToneCurve.BuildTabulated16(Context, count, null);
+                newGamma = ToneCurve.BuildTabulated16(Context, (int)count, null);
                 if (newGamma is null) return null;
 
                 if (!io.ReadUInt16Array((int)count, out newGamma.Table16)) {
